fix css for break line
